ajout jackson get periodeTravailles get rapportVisites get user revois seulements les données relatif a l'utilisateur courant ajout de usernotfound exception ajout de methode dans les repo get user throws exeption si pas de user ou invalid token ... ajout dune strategie de serialisation avec jackson dans view et anotation des entity ajout de create response dans baseController

+++ src/main/java/LPY/appliVisiteur/Model/Repository/UserRepository.java
@@ -7,5 +7,5 @@ import java.util.Optional;
 
 public interface UserRepository extends CrudRepository<User, Integer> {
     User findByLogin(String login);
-    Optional<User> findById(Integer id);
+    Optional<User> findById(long id);
 }
